Change type for body to LargeBinaryArray.
When migrating lots of huge private warc files, I noticed that it's possible for a chunk to exceed u32 memory limit of BinaryArray within the default batch size. Instead of changing the batch size, I decided to change typing - hopefully this makes the project more stable in this specific case.